requirements.txt
discord.py==1.7.3
plotly==5.1.0
requests==2.25.1
websockets==9.1
selenium==3.141.0
git+git://github.com/Rapptz/asqlite
bar_chart_race==0.1.0
blend_modes==2.1.0
opencv_python==4.5.3.56
numpy==1.20.3
pandas==1.3.0
discord_py_interactions==3.0.2
pytz==2021.1
matplotlib==3.4.2
aiohttp==3.7.4.post0
alive_progress==1.6.2
beautifulsoup4==4.10.0
discord==1.7.3
Pillow==8.3.2
python-dotenv==0.19.0
kaleido==0.2.1
